+++ packages/core/.storybook/main.js
@@ -1,6 +1,9 @@
 module.exports = {
   stories: ['../stories/*.stories.tsx'],
   logLevel: 'debug',
+  core: {
+    builder: 'webpack5',
+  },
   addons: [
     {
       name: '@storybook/addon-essentials',
@@ -22,4 +25,23 @@ module.exports = {
     },
     './src/theme-changer/register',
   ],
+  webpackFinal: async config => {
+    // replace asset loader with url loader in order to bundle icons into js
+    function isAssetLoader(rule) {
+      return rule.test.toString().includes('png');
+    }
+
+    const assetLoader = config.module.rules.find(isAssetLoader);
+
+    config.module.rules = config.module.rules.filter(
+      rule => !isAssetLoader(rule),
+    );
+
+    config.module.rules.push({
+      test: assetLoader.test,
+      loader: 'url-loader',
+    });
+
+    return config;
+  },
 };