+++ packages/gatsby-theme/src/components/seo.js
@@ -1,0 +1,53 @@
+import React from 'react';
+import { Helmet } from 'react-helmet';
+import { useStaticQuery, graphql } from 'gatsby';
+
+const SEO = ({ content }) => {
+  const data = useStaticQuery(graphql`
+    {
+      site {
+        siteMetadata {
+          title
+          description
+          siteUrl
+          image
+        }
+      }
+    }
+  `);
+
+  const defaults = data.site.siteMetadata;
+
+  const title = content.frontmatter?.title || defaults.title;
+  const description = content.frontmatter?.description || defaults.description;
+  const url = `${defaults.siteUrl}/${content.slug || ''}`;
+  const image =
+    content.frontmatter?.image || defaults.image
+      ? `${defaults.siteUrl}${content.frontmatter?.image || defaults.image}`
+      : false;
+  const author = content.frontmatter?.author || defaults.author;
+
+  return (
+    <>
+      <Helmet>
+        <title>{title}</title>
+        <link rel="canonical" href={url} />
+        <meta name="description" content={description} />
+        {image && <meta name="image" content={image} />}
+
+        <meta property="og:url" content={url} />
+        <meta property="og:title" content={title} />
+        <meta property="og:description" content={description} />
+        {image && <meta property="og:image" content={image} />}
+
+        <meta name="twitter:card" content="summary_large_image" />
+        <meta name="twitter:creator" content={author} />
+        <meta name="twitter:title" content={title} />
+        <meta name="twitter:description" content={description} />
+        {image && <meta name="twitter:image" content={image} />}
+      </Helmet>
+    </>
+  );
+};
+
+export default SEO;