+++ packages/gatsby-theme/src/utils/index.js
@@ -1,0 +1,5 @@
+import isEmpty from './isEmpty';
+import navify from './navify';
+import isMobile from './isMobile';
+
+export { isEmpty, isMobile, navify };